+++ pkgs/native_assets_builder/lib/native_assets_builder.dart
@@ -3,4 +3,5 @@
 // BSD-style license that can be found in the LICENSE file.
 
 export 'package:native_assets_builder/src/build_runner/build_runner.dart';
+export 'package:native_assets_builder/src/model/asset.dart';
 export 'package:native_assets_builder/src/package_layout/package_layout.dart';